prover(refactor): renaming to remove artifacts
